pyproject.toml
[tool.poetry]
name = "blacklistusers"
version = "0.1.0"
description = ""
authors = ["AndreyKosov <andrey.kosov.05@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.5"}
aiogram = "^3.15.0"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
